# src/main/java/com/dreptschar/plantmanagerbackend/adapter/inbound/rest/api/ActionPlansApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.13.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.dreptschar.plantmanagerbackend.adapter.inbound.rest.api;

import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.annotation.Generated;
import jakarta.validation.constraints.*;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-07-13T15:35:07.946460567Z[Etc/UTC]", comments = "Generator version: 7.13.0")
@Validated
@Tag(name = "Action Plans", description = "the Action Plans API")
public interface ActionPlansApi {

    /**
     * GET /plants/{plantId}/action-plans : List action plans for a plant
     *
     * @param plantId  (required)
     * @return Action plan list (status code 200)
     */
    @Operation(
        operationId = "plantsPlantIdActionPlansGet",
        summary = "List action plans for a plant",
        tags = { "Action Plans" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Action plan list")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/plants/{plantId}/action-plans"
    )
    
    default ResponseEntity<Void> plantsPlantIdActionPlansGet(
        @Parameter(name = "plantId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("plantId") Long plantId
    ) {
        return ResponseEntity.status(HttpStatus.NOT_IMPLEMENTED).build();
    }


    /**
     * POST /plants/{plantId}/action-plans : Create an action plan for a plant
     *
     * @param plantId  (required)
     * @return Action plan created (status code 201)
     */
    @Operation(
        operationId = "plantsPlantIdActionPlansPost",
        summary = "Create an action plan for a plant",
        tags = { "Action Plans" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Action plan created")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/plants/{plantId}/action-plans"
    )
    
    default ResponseEntity<Void> plantsPlantIdActionPlansPost(
        @Parameter(name = "plantId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("plantId") Long plantId
    ) {
        return ResponseEntity.status(HttpStatus.NOT_IMPLEMENTED).build();
    }

}