# src/main/java/com/dreptschar/plantmanagerbackend/adapter/inbound/rest/api/PlantActionsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.13.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.dreptschar.plantmanagerbackend.adapter.inbound.rest.api;

import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.annotation.Generated;
import jakarta.validation.constraints.*;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-07-13T15:35:07.946460567Z[Etc/UTC]", comments = "Generator version: 7.13.0")
@Validated
@Tag(name = "Plant Actions", description = "the Plant Actions API")
public interface PlantActionsApi {

    /**
     * GET /plants/{plantId}/actions : List actions performed on a plant
     *
     * @param plantId  (required)
     * @return Action history (status code 200)
     */
    @Operation(
        operationId = "plantsPlantIdActionsGet",
        summary = "List actions performed on a plant",
        tags = { "Plant Actions" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Action history")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/plants/{plantId}/actions"
    )
    
    default ResponseEntity<Void> plantsPlantIdActionsGet(
        @Parameter(name = "plantId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("plantId") Long plantId
    ) {
        return ResponseEntity.status(HttpStatus.NOT_IMPLEMENTED).build();
    }


    /**
     * POST /plants/{plantId}/actions : Record an action for a plant
     *
     * @param plantId  (required)
     * @return Action recorded (status code 201)
     */
    @Operation(
        operationId = "plantsPlantIdActionsPost",
        summary = "Record an action for a plant",
        tags = { "Plant Actions" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Action recorded")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/plants/{plantId}/actions"
    )
    
    default ResponseEntity<Void> plantsPlantIdActionsPost(
        @Parameter(name = "plantId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("plantId") Long plantId
    ) {
        return ResponseEntity.status(HttpStatus.NOT_IMPLEMENTED).build();
    }

}