# src/main/java/com/dreptschar/plantmanagerbackend/adapter/inbound/rest/api/PlantsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.13.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.dreptschar.plantmanagerbackend.adapter.inbound.rest.api;

import com.dreptschar.plantmanagerbackend.adapter.inbound.rest.api.model.PlantSummaryResponse;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.annotation.Generated;
import java.util.List;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-07-13T15:35:07.946460567Z[Etc/UTC]", comments = "Generator version: 7.13.0")
@Validated
@Tag(name = "Plants", description = "the Plants API")
public interface PlantsApi {

    /**
     * GET /plants : List plants
     *
     * @return Plant list (status code 200)
     */
    @Operation(
        operationId = "plantsGet",
        summary = "List plants",
        tags = { "Plants" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Plant list")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/plants"
    )
    
    default ResponseEntity<List<PlantSummaryResponse>> plantsGet() {
        return ResponseEntity.status(HttpStatus.NOT_IMPLEMENTED).build();
    }


    /**
     * DELETE /plants/{plantId} : Delete a plant
     *
     * @param plantId  (required)
     * @return Plant deleted (status code 204)
     *         or Not found (status code 404)
     */
    @Operation(
        operationId = "plantsPlantIdDelete",
        summary = "Delete a plant",
        tags = { "Plants" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Plant deleted"),
            @ApiResponse(responseCode = "404", description = "Not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/plants/{plantId}"
    )
    
    default ResponseEntity<Void> plantsPlantIdDelete(
        @Parameter(name = "plantId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("plantId") Long plantId
    ) {
        return ResponseEntity.status(HttpStatus.NOT_IMPLEMENTED).build();
    }


    /**
     * GET /plants/{plantId} : Get a plant by id
     *
     * @param plantId  (required)
     * @return Plant (status code 200)
     *         or Not found (status code 404)
     */
    @Operation(
        operationId = "plantsPlantIdGet",
        summary = "Get a plant by id",
        tags = { "Plants" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Plant"),
            @ApiResponse(responseCode = "404", description = "Not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/plants/{plantId}"
    )
    
    default ResponseEntity<Void> plantsPlantIdGet(
        @Parameter(name = "plantId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("plantId") Long plantId
    ) {
        return ResponseEntity.status(HttpStatus.NOT_IMPLEMENTED).build();
    }


    /**
     * PATCH /plants/{plantId} : Update a plant
     *
     * @param plantId  (required)
     * @return Plant updated (status code 200)
     *         or Not found (status code 404)
     */
    @Operation(
        operationId = "plantsPlantIdPatch",
        summary = "Update a plant",
        tags = { "Plants" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Plant updated"),
            @ApiResponse(responseCode = "404", description = "Not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.PATCH,
        value = "/plants/{plantId}"
    )
    
    default ResponseEntity<Void> plantsPlantIdPatch(
        @Parameter(name = "plantId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("plantId") Long plantId
    ) {
        return ResponseEntity.status(HttpStatus.NOT_IMPLEMENTED).build();
    }


    /**
     * POST /plants : Create a plant
     *
     * @return Plant created (status code 201)
     */
    @Operation(
        operationId = "plantsPost",
        summary = "Create a plant",
        tags = { "Plants" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Plant created")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/plants"
    )
    
    default ResponseEntity<Void> plantsPost(
        
    ) {
        return ResponseEntity.status(HttpStatus.NOT_IMPLEMENTED).build();
    }

}