# src/main/java/com/dreptschar/plantmanagerbackend/adapter/inbound/rest/api/RoomsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.13.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.dreptschar.plantmanagerbackend.adapter.inbound.rest.api;

import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.tags.Tag;
import jakarta.annotation.Generated;
import jakarta.validation.constraints.*;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-07-13T15:35:07.946460567Z[Etc/UTC]", comments = "Generator version: 7.13.0")
@Validated
@Tag(name = "Rooms", description = "the Rooms API")
public interface RoomsApi {

    /**
     * GET /rooms : List rooms
     *
     * @return Room list (status code 200)
     */
    @Operation(
        operationId = "roomsGet",
        summary = "List rooms",
        tags = { "Rooms" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Room list")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/rooms"
    )
    
    default ResponseEntity<Void> roomsGet(
        
    ) {
        return ResponseEntity.status(HttpStatus.NOT_IMPLEMENTED).build();
    }


    /**
     * POST /rooms : Create a room
     *
     * @return Room created (status code 201)
     */
    @Operation(
        operationId = "roomsPost",
        summary = "Create a room",
        tags = { "Rooms" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Room created")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/rooms"
    )
    
    default ResponseEntity<Void> roomsPost(
        
    ) {
        return ResponseEntity.status(HttpStatus.NOT_IMPLEMENTED).build();
    }


    /**
     * DELETE /rooms/{roomId} : Delete a room
     *
     * @param roomId  (required)
     * @return Room deleted (status code 204)
     *         or Not found (status code 404)
     */
    @Operation(
        operationId = "roomsRoomIdDelete",
        summary = "Delete a room",
        tags = { "Rooms" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Room deleted"),
            @ApiResponse(responseCode = "404", description = "Not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/rooms/{roomId}"
    )
    
    default ResponseEntity<Void> roomsRoomIdDelete(
        @Parameter(name = "roomId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("roomId") Long roomId
    ) {
        return ResponseEntity.status(HttpStatus.NOT_IMPLEMENTED).build();
    }


    /**
     * GET /rooms/{roomId} : Get a room by id
     *
     * @param roomId  (required)
     * @return Room (status code 200)
     *         or Not found (status code 404)
     */
    @Operation(
        operationId = "roomsRoomIdGet",
        summary = "Get a room by id",
        tags = { "Rooms" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Room"),
            @ApiResponse(responseCode = "404", description = "Not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/rooms/{roomId}"
    )
    
    default ResponseEntity<Void> roomsRoomIdGet(
        @Parameter(name = "roomId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("roomId") Long roomId
    ) {
        return ResponseEntity.status(HttpStatus.NOT_IMPLEMENTED).build();
    }


    /**
     * PATCH /rooms/{roomId} : Update a room
     *
     * @param roomId  (required)
     * @return Room updated (status code 200)
     *         or Not found (status code 404)
     */
    @Operation(
        operationId = "roomsRoomIdPatch",
        summary = "Update a room",
        tags = { "Rooms" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Room updated"),
            @ApiResponse(responseCode = "404", description = "Not found")
        }
    )
    @RequestMapping(
        method = RequestMethod.PATCH,
        value = "/rooms/{roomId}"
    )
    
    default ResponseEntity<Void> roomsRoomIdPatch(
        @Parameter(name = "roomId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("roomId") Long roomId
    ) {
        return ResponseEntity.status(HttpStatus.NOT_IMPLEMENTED).build();
    }

}