# src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  title: PlantManagerBackend API
  version: 0.1.0
  description: REST API for managing house plants, recurring care plans, and action history.

servers:
  - url: /api
    description: Default API base path

tags:
  - name: Plants
  - name: Rooms
  - name: Action Types
  - name: Action Plans
  - name: Plant Actions

paths:
  /rooms:
    get:
      tags: [Rooms]
      summary: List rooms
      responses:
        "200":
          description: Room list
    post:
      tags: [Rooms]
      summary: Create a room
      responses:
        "201":
          description: Room created

  /rooms/{roomId}:
    get:
      tags: [Rooms]
      summary: Get a room by id
      parameters:
        - $ref: "#/components/parameters/RoomId"
      responses:
        "200":
          description: Room
        "404":
          description: Not found
    patch:
      tags: [Rooms]
      summary: Update a room
      parameters:
        - $ref: "#/components/parameters/RoomId"
      responses:
        "200":
          description: Room updated
        "404":
          description: Not found
    delete:
      tags: [Rooms]
      summary: Delete a room
      parameters:
        - $ref: "#/components/parameters/RoomId"
      responses:
        "204":
          description: Room deleted
        "404":
          description: Not found

  /plants:
    get:
      tags: [Plants]
      summary: List plants
      responses:
        "200":
          description: Plant list
    post:
      tags: [Plants]
      summary: Create a plant
      responses:
        "201":
          description: Plant created

  /plants/{plantId}:
    get:
      tags: [Plants]
      summary: Get a plant by id
      parameters:
        - $ref: "#/components/parameters/PlantId"
      responses:
        "200":
          description: Plant
        "404":
          description: Not found
    patch:
      tags: [Plants]
      summary: Update a plant
      parameters:
        - $ref: "#/components/parameters/PlantId"
      responses:
        "200":
          description: Plant updated
        "404":
          description: Not found
    delete:
      tags: [Plants]
      summary: Delete a plant
      parameters:
        - $ref: "#/components/parameters/PlantId"
      responses:
        "204":
          description: Plant deleted
        "404":
          description: Not found

  /plants/{plantId}/actions:
    get:
      tags: [Plant Actions]
      summary: List actions performed on a plant
      parameters:
        - $ref: "#/components/parameters/PlantId"
      responses:
        "200":
          description: Action history
    post:
      tags: [Plant Actions]
      summary: Record an action for a plant
      parameters:
        - $ref: "#/components/parameters/PlantId"
      responses:
        "201":
          description: Action recorded

  /plants/{plantId}/action-plans:
    get:
      tags: [Action Plans]
      summary: List action plans for a plant
      parameters:
        - $ref: "#/components/parameters/PlantId"
      responses:
        "200":
          description: Action plan list
    post:
      tags: [Action Plans]
      summary: Create an action plan for a plant
      parameters:
        - $ref: "#/components/parameters/PlantId"
      responses:
        "201":
          description: Action plan created

  /action-types:
    get:
      tags: [Action Types]
      summary: List action types
      responses:
        "200":
          description: Action type list

components:
  parameters:
    PlantId:
      name: plantId
      in: path
      required: true
      schema:
        type: integer
        format: int64
    RoomId:
      name: roomId
      in: path
      required: true
      schema:
        type: integer
        format: int64